Correção da implementação do método criar de personagem
