+++ Cod3rsGrowth.Service/IPersonagemServico.cs
@@ -1,21 +1,11 @@
-using CodersGrowth.Domain;
+using CodersGrowth.Domain.Entities;
 
 namespace Cod3rsGrowth.Service
 {
     public interface IPersonagemServico
     {
-        // public List<Personagem> ObterTodos(string filtro);
-        // public bool ValidarObterTodos(string filtro);
-
-        // public Personagem ObterPorId(int id);
-        // public bool ValidarObterPorId(int id);
-
+        public List<Personagem> ObterTodos();
+        public Personagem? ObterPorId(int id);
         public int Criar(Personagem personagem);
-
-        // public void Editar(int id);
-        // public bool ValidarEditar(int id);
-
-        // public void Remover(int id);
-        // public bool ValidarRemover(int id);
     }
 }
