+++ Cod3rsGrowth.Tests/Tests/test_personagem_servico.cs
@@ -1,6 +1,7 @@
 using Cod3rsGrowth.Infra;
 using Cod3rsGrowth.Service;
-using CodersGrowth.Domain;
+using CodersGrowth.Domain.Enums;
+using CodersGrowth.Domain.Entities;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Cod3rsGrowth.Tests.Tests
@@ -8,27 +9,25 @@ namespace Cod3rsGrowth.Tests.Tests
     public class TestePersonagemServicos : TesteBase
     {
         private readonly IPersonagemServico personagemService;
-        private readonly IPersonagemRepositorio personagemRepositorio;
 
         public TestePersonagemServicos() : base()
         {
             personagemService = ServiceProvider.GetRequiredService<IPersonagemServico>();
-            personagemRepositorio = ServiceProvider.GetRequiredService<IPersonagemRepositorio>();
         }
 
         [Fact]
-        public void TestarObterTodos()
+        public void TestarCriarNovoPersonagem()
         {
              // Arrange
-            var novoPersonagem = new Personagem("NovoPersonagem", 100, 50, 1.0f, CategoriasEnum.Bom, CategoriasEnum.Medio);
+            var novoPersonagem = new Personagem(null, "Teste da Silva", 100, 50, 1.0f, CategoriasEnum.Bom, CategoriasEnum.Medio);
 
             // Act
             var idCriado = personagemService.Criar(novoPersonagem);
+            var personagemCriado = personagemService.ObterPorId(idCriado);
 
             // Assert
-            Assert.Equal(9, idCriado);
-            var personagemCriado = personagemRepositorio.ObterTodos("NovoPersonagem").FirstOrDefault();
             Assert.NotNull(personagemCriado);
+            Assert.Equal(personagemCriado.Id, idCriado);
             Assert.Equal(novoPersonagem.Nome, personagemCriado.Nome);
         }
     }
